+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Prevent setup to run again if already configured when using `bake`
 - Fix `Unique::getPaginatedQuery` to call to `addSelect` instead of `select` during the pre-paginated query
 - Updated Spanish Translation (Thanks @silvioq !)
+- Fix error template in WhoopsRenderer (#885; Thanks @silvioq !)
 
 ## v4.1.18
 - Bakery setup wizard for SMTP config + separate SMTP setup in it's own command (https://github.com/userfrosting/UserFrosting/issues/874)